+++ libvector/include/libvector.h
@@ -49,7 +49,7 @@ typedef struct		s_ivec2
 */
 
 t_vec3 		vec_diff(t_vec3 v1, t_vec3 v2);
-double		dot_product(t_vec3 v1, t_vec3 v2);
+float		dot_product(t_vec3 v1, t_vec3 v2);
 t_vec3 		vec_add(t_vec3 v1, t_vec3 v2);
 t_vec3		vec_normalize(t_vec3 v);
 t_vec3		mult_vec_const(t_vec3 v, float n);
